+++ lib/Lazysite/Manager/Backups.pm
@@ -18,7 +18,8 @@ use Lazysite::Util    qw(log_event);
 use Exporter          qw(import);
 our @EXPORT_OK = qw(action_backup_list action_backup_create action_backup_download
     action_backup_restore action_backup_delete
-    write_sha256 read_sha256 verify_sha256);
+    write_sha256 read_sha256 verify_sha256
+    _claim_name);
 
 our $DOCROOT      = '';
 our $LAZYSITE_DIR = '';